+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 
+### Added
+
+* New tap action: choose the next photo
+
 ## [v1.28.4] - 2025-04-09
 [v1.28.4]: https://github.com/fibelatti/photo-widget/releases/tag/v1.28.4